refactor(trigger-api): extract server into internal/triggerapi
