+++ internal/triggerapi/routes.go
@@ -1,0 +1,25 @@
+package triggerapi
+
+import (
+	"encoding/json"
+	"net/http"
+)
+
+func Routes(s *Server) http.Handler {
+	mux := http.NewServeMux()
+	mux.HandleFunc("GET /healthz", func(w http.ResponseWriter, r *http.Request) {
+		writeJSON(w, http.StatusOK, map[string]any{"ok": true})
+	})
+	mux.HandleFunc("POST /v1/events/gitea", s.handleGitea)
+	mux.HandleFunc("POST /v1/tasks", s.handleManualTask)
+	mux.HandleFunc("GET /v1/tasks", s.handleListTasks)
+	mux.HandleFunc("GET /v1/tasks/{id}", s.handleGetTask)
+	mux.HandleFunc("GET /v1/tasks/{id}/events", s.handleTaskEvents)
+	return mux
+}
+
+func writeJSON(w http.ResponseWriter, code int, v any) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(code)
+	_ = json.NewEncoder(w).Encode(v)
+}
